pyproject.toml
[project]
name = "generating-shakespearean-text-using-char-rnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "e>=1.4.5",
    "numpy>=2.4.0",
    "tokenizer>=3.6.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ruff>=0.14.10",
]
